requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
colorama==0.4.3
coverage==5.3
cryptography==3.1
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.0.2
pytest-cov==2.10.1
pytest-flask==1.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-telegram-bot==12.8
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
tornado==6.0.4
Werkzeug==1.0.1
